feat: adds slog support
Now the Bootstrap function also configures a slog handler that uses the
zerolog implementation and the same configuration as defined in the
-app-log-* flags.

This fixes log levels when logging happens using the official slog
package. This is normally done by frameworks and other similar packages.

This happens automatically and will work out of the box.

This commits also updates all dependencies.