feat: self probe check (#34)
Dockerfile and docker-compose both have specific configuration for
checking the health of a container. go-app already provides an URL for
probing, but the container had to have wget o curl por probing.

This commit introduces a two new configuration that allows for the
binary to check it's own probes.

if one of -app-check-ready or -app-check-healthy is set the binary will
try to check http://localhost:port/ready or /healthy instead of
bootstrap the new app.

Links: https://docs.docker.com/reference/dockerfile/#healthcheck